+++ src/ts/audiotrackutils.ts
@@ -1,5 +1,6 @@
 import {ListSelector, ListSelectorConfig} from './components/listselector';
-import { InternalUIConfig, UIInstanceManager, UIManager } from './uimanager';
+import { UIInstanceManager } from './uimanager';
+import { PlayerAPI } from 'bitmovin-player';
 
 /**
  * Helper class to handle all audio tracks related events
@@ -8,11 +9,11 @@ import { InternalUIConfig, UIInstanceManager, UIManager } from './uimanager';
  */
 export class AudioTrackSwitchHandler {
 
-  private player: bitmovin.PlayerAPI;
+  private player: PlayerAPI;
   private listElement: ListSelector<ListSelectorConfig>;
   private uimanager: UIInstanceManager;
 
-  constructor(player: bitmovin.PlayerAPI, element: ListSelector<ListSelectorConfig>, uimanager: UIInstanceManager) {
+  constructor(player: PlayerAPI, element: ListSelector<ListSelectorConfig>, uimanager: UIInstanceManager) {
     this.player = player;
     this.listElement = element;
     this.uimanager = uimanager;
@@ -31,16 +32,16 @@ export class AudioTrackSwitchHandler {
   private bindPlayerEvents(): void {
     const updateAudioTracksCallback = (): void => this.updateAudioTracks();
     // Update selection when selected track has changed
-    this.player.on(this.player.exports.Event.AudioChanged, () => {
+    this.player.on(this.player.exports.PlayerEvent.AudioChanged, () => {
       this.selectCurrentAudioTrack();
     });
     // Update tracks when source goes away
-    this.player.on(this.player.exports.Event.SourceUnloaded, updateAudioTracksCallback);
+    this.player.on(this.player.exports.PlayerEvent.SourceUnloaded, updateAudioTracksCallback);
     // Update tracks when the period within a source changes
-    this.player.on(this.player.exports.Event.PeriodSwitched, updateAudioTracksCallback);
+    this.player.on(this.player.exports.PlayerEvent.PeriodSwitched, updateAudioTracksCallback);
     // Update tracks when a track is added or removed
-    this.player.on(this.player.exports.Event.AudioAdded, updateAudioTracksCallback);
-    this.player.on(this.player.exports.Event.AudioRemoved, updateAudioTracksCallback);
+    this.player.on(this.player.exports.PlayerEvent.AudioAdded, updateAudioTracksCallback);
+    this.player.on(this.player.exports.PlayerEvent.AudioRemoved, updateAudioTracksCallback);
     this.uimanager.getConfig().events.onUpdated.subscribe(updateAudioTracksCallback);
   }