+++ src/ts/components/audiotrackselectbox.ts
@@ -2,6 +2,7 @@ import {SelectBox} from './selectbox';
 import {ListSelectorConfig} from './listselector';
 import {UIInstanceManager} from '../uimanager';
 import {AudioTrackSwitchHandler} from '../audiotrackutils';
+import { PlayerAPI } from 'bitmovin-player';
 
 /**
  * A select box providing a selection between available audio tracks (e.g. different languages).
@@ -12,7 +13,7 @@ export class AudioTrackSelectBox extends SelectBox {
     super(config);
   }
 
-  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
+  configure(player: PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
     new AudioTrackSwitchHandler(player, this, uimanager);