+++ src/ts/components/playbacktogglebutton.ts
@@ -1,9 +1,8 @@
 import {ToggleButton, ToggleButtonConfig} from './togglebutton';
 import {UIInstanceManager} from '../uimanager';
-import PlayerEvent = bitmovin.PlayerAPI.PlayerEvent;
 import {PlayerUtils} from '../playerutils';
 import TimeShiftAvailabilityChangedArgs = PlayerUtils.TimeShiftAvailabilityChangedArgs;
-import WarningEvent = bitmovin.PlayerAPI.WarningEvent;
+import { PlayerAPI, WarningEvent } from 'bitmovin-player';
 
 /**
  * A button that toggles between playback and pause.
@@ -24,7 +23,7 @@ export class PlaybackToggleButton extends ToggleButton<ToggleButtonConfig> {
     this.isPlayInitiated = false;
   }
 
-  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager, handleClickEvent: boolean = true): void {
+  configure(player: PlayerAPI, uimanager: UIInstanceManager, handleClickEvent: boolean = true): void {
     super.configure(player, uimanager);
 
     let isSeeking = false;
@@ -45,32 +44,32 @@ export class PlaybackToggleButton extends ToggleButton<ToggleButtonConfig> {
     };
 
     // Call handler upon these events
-    player.on(player.exports.Event.Play, (e) => {
+    player.on(player.exports.PlayerEvent.Play, (e) => {
       this.isPlayInitiated = true;
       playbackStateHandler();
     });
 
-    player.on(player.exports.Event.Paused, (e) => {
+    player.on(player.exports.PlayerEvent.Paused, (e) => {
       this.isPlayInitiated = false;
       playbackStateHandler();
     });
 
-    player.on(player.exports.Event.Playing, (e) => {
+    player.on(player.exports.PlayerEvent.Playing, (e) => {
       this.isPlayInitiated = false;
       playbackStateHandler();
     });
     // after unloading + loading a new source, the player might be in a different playing state (from playing into stopped)
-    player.on(player.exports.Event.SourceLoaded, playbackStateHandler);
+    player.on(player.exports.PlayerEvent.SourceLoaded, playbackStateHandler);
     uimanager.getConfig().events.onUpdated.subscribe(playbackStateHandler);
-    player.on(player.exports.Event.SourceUnloaded, playbackStateHandler);
+    player.on(player.exports.PlayerEvent.SourceUnloaded, playbackStateHandler);
     // when playback finishes, player turns to paused mode
-    player.on(player.exports.Event.PlaybackFinished, playbackStateHandler);
-    player.on(player.exports.Event.CastStarted, playbackStateHandler);
+    player.on(player.exports.PlayerEvent.PlaybackFinished, playbackStateHandler);
+    player.on(player.exports.PlayerEvent.CastStarted, playbackStateHandler);
 
     // When a playback attempt is rejected with warning 5008, we switch the button state back to off
     // This is required for blocked autoplay, because there is no Paused event in such case
-    player.on(player.exports.Event.Warning, (event: WarningEvent) => {
-      if (event.code === 5008) {
+    player.on(player.exports.PlayerEvent.Warning, (event: WarningEvent) => {
+      if (event.code === player.exports.WarningCode.PLAYBACK_COULD_NOT_BE_STARTED) {
         this.isPlayInitiated = false;
         this.off();
       }