+++ src/ts/components/settingspanelpagebackbutton.ts
@@ -1,5 +1,6 @@
 import {UIInstanceManager} from '../uimanager';
 import {SettingsPanelPageNavigatorButton, SettingsPanelPageNavigatorConfig} from './settingspanelpagenavigatorbutton';
+import { PlayerAPI } from 'bitmovin-player';
 
 export class SettingsPanelPageBackButton extends SettingsPanelPageNavigatorButton {
 
@@ -12,7 +13,7 @@ export class SettingsPanelPageBackButton extends SettingsPanelPageNavigatorButto
     }, this.config);
   }
 
-  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
+  configure(player: PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager);
 
     this.onClick.subscribe(() => {