+++ src/ts/components/volumeslider.ts
@@ -1,5 +1,6 @@
 import {SeekBar, SeekBarConfig} from './seekbar';
 import {UIInstanceManager} from '../uimanager';
+import { PlayerAPI } from 'bitmovin-player';
 
 /**
  * Configuration interface for the {@link VolumeSlider} component.
@@ -29,7 +30,7 @@ export class VolumeSlider extends SeekBar {
     }, this.config);
   }
 
-  configure(player: bitmovin.PlayerAPI, uimanager: UIInstanceManager): void {
+  configure(player: PlayerAPI, uimanager: UIInstanceManager): void {
     super.configure(player, uimanager, false);
 
     let config = <VolumeSliderConfig>this.getConfig();
@@ -50,9 +51,9 @@ export class VolumeSlider extends SeekBar {
       }
     };
 
-    player.on(player.exports.Event.VolumeChanged, volumeChangeHandler);
-    player.on(player.exports.Event.Muted, volumeChangeHandler);
-    player.on(player.exports.Event.Unmuted, volumeChangeHandler);
+    player.on(player.exports.PlayerEvent.VolumeChanged, volumeChangeHandler);
+    player.on(player.exports.PlayerEvent.Muted, volumeChangeHandler);
+    player.on(player.exports.PlayerEvent.Unmuted, volumeChangeHandler);
     uimanager.getConfig().events.onUpdated.subscribe(volumeChangeHandler);
 
     this.onSeekPreview.subscribeRateLimited((sender, args) => {
@@ -66,7 +67,7 @@ export class VolumeSlider extends SeekBar {
 
     // Update the volume slider marker when the player resized, a source is loaded,
     // or the UI is configured. Check the seekbar for a detailed description.
-    player.on(player.exports.Event.PlayerResized, () => {
+    player.on(player.exports.PlayerEvent.PlayerResized, () => {
       this.refreshPlaybackPosition();
     });
     uimanager.onConfigured.subscribe(() => {