+++ src/ts/playerutils.ts
@@ -1,10 +1,9 @@
 import {Event, EventDispatcher, NoArgs} from './eventdispatcher';
 import {BrowserUtils} from './browserutils';
-import { InternalUIConfig, UIInstanceManager, UIManager } from './uimanager';
+import { UIInstanceManager } from './uimanager';
+import { PlayerAPI } from 'bitmovin-player';
 
 export namespace PlayerUtils {
-
-  import PlayerAPI = bitmovin.PlayerAPI;
 
   export enum PlayerState {
     Idle,
@@ -14,7 +13,7 @@ export namespace PlayerUtils {
     Finished,
   }
 
-  export function isTimeShiftAvailable(player: bitmovin.PlayerAPI): boolean {
+  export function isTimeShiftAvailable(player: PlayerAPI): boolean {
     return player.isLive() && player.getMaxTimeShift() !== 0;
   }
 
@@ -50,10 +49,10 @@ export namespace PlayerUtils {
         this.detect();
       };
       // Try to detect timeshift availability when source is loaded, which works for DASH streams
-      player.on(player.exports.Event.SourceLoaded, timeShiftDetector);
+      player.on(player.exports.PlayerEvent.SourceLoaded, timeShiftDetector);
       // With HLS/NativePlayer streams, getMaxTimeShift can be 0 before the buffer fills, so we need to additionally
       // check timeshift availability in TimeChanged
-      player.on(player.exports.Event.TimeChanged, timeShiftDetector);
+      player.on(player.exports.PlayerEvent.TimeChanged, timeShiftDetector);
     }
 
     detect(): void {
@@ -106,13 +105,13 @@ export namespace PlayerUtils {
       };
       this.uimanager.getConfig().events.onUpdated.subscribe(liveDetector);
       // Re-evaluate when playback starts
-      player.on(player.exports.Event.Play, liveDetector);
+      player.on(player.exports.PlayerEvent.Play, liveDetector);
 
       // HLS live detection workaround for Android:
       // Also re-evaluate during playback, because that is when the live flag might change.
       // (Doing it only in Android Chrome saves unnecessary overhead on other plattforms)
       if (BrowserUtils.isAndroid && BrowserUtils.isChrome) {
-        player.on(player.exports.Event.TimeChanged, liveDetector);
+        player.on(player.exports.PlayerEvent.TimeChanged, liveDetector);
       }
     }